Rename ring/ to orbis/ and fix process leak in fixture
- Rename e2e module from ring/ to orbis/ (matches project name)
- Replace static OnceCell fixture with owned setup_dkg() that returns
  DkgFixture by value — processes are now properly cleaned up on drop
- Remove shared_dkg_fixture() since we have a single integration test
